NanRenHu/requirements.txt
SQLAlchemy==1.2.15
Flask==1.0.2
beautifulsoup4==4.8.1
python_dateutil==2.8.0
paho_mqtt==1.4.0
